task9/requirements.txt
Django==2.0.2
mongoengine==0.15.0
Pillow==5.0.0
djangorestframework
django-rest-framework-mongoengine
django-allauth
django-cors-headers
django-bootstrap4==0.0.6
django-serializer
djangorestframework-jwt
